+++ src/scripts/helpers/blockChecksArray.js
@@ -1,0 +1,3 @@
+import { checkHeadingLevel } from '../blockChecks/checkHeading'
+
+export const blockChecksArray = [checkHeadingLevel]