+++ src/scripts/helpers/getInvalidBlocks.js
@@ -1,0 +1,9 @@
+import { useSelect } from '@wordpress/data'
+import { blockChecksArray } from './blockChecksArray'
+
+export function getInvalidBlocks() {
+  const allBlocks = useSelect((select) => select('core/block-editor').getBlocks(), [])
+  return allBlocks
+    .flatMap((block) => blockChecksArray.map((check) => check(block)))
+    .filter((result) => !result.isValid)
+}